Move lots of menu code over to menu/menu_driver.c - moved at least 44Kb of code over
